refactor: modularize playwright stub runner
